# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ultiproxy Public API
  version: 1.0.0
  description: |
    Ultiproxy is an agent-first universal LLM subscription proxy.
    It exposes a single OpenAI-compatible and Anthropic-compatible gateway at port 9050,
    fronting multiple pooled subscriptions (GitHub Copilot, OpenAI Codex, Anthropic Claude,
    Google Antigravity, xAI Grok, DeepSeek, Z.ai, local vLLM, etc.) with live quota monitoring,
    per-API-key accounting, and an embedded Model Context Protocol (MCP) server.
servers:
  - url: http://localhost:9050
    description: Local Ultiproxy Gateway Daemon
  - url: http://127.0.0.1:9050
    description: Loopback address
security:
  # Either spelling is accepted on every authenticated route: the OpenAI-style
  # bearer header, or the Anthropic-style x-api-key header.
  - bearerAuth: []
  - apiKeyHeader: []
paths:
  /v1/chat/completions:
    post:
      summary: Create chat completion
      description: |
        OpenAI-compatible chat completion endpoint supporting multi-turn conversations,
        streaming Server-Sent Events (SSE), tool calling / function calling, vision/image inputs,
        and reasoning deltas (o1/DeepSeek-R1 style).
      operationId: createChatCompletion
      tags:
        - OpenAI Compatibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Successful completion or streaming event stream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
                description: Real-time Server-Sent Events (SSE) data stream of ChatCompletionChunk objects.
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/BadGatewayError'
  /v1/messages:
    post:
      summary: Create Anthropic message
      description: |
        Anthropic-compatible Messages API endpoint for direct use by Claude Code, Hermes,
        and Anthropic SDKs. Supports system prompts, structured content blocks, and streaming.
      operationId: createAnthropicMessage
      tags:
        - Anthropic Compatibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnthropicMessageRequest'
      responses:
        '200':
          description: Anthropic message response or event stream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnthropicMessageResponse'
            text/event-stream:
              schema:
                type: string
                description: Anthropic SSE message stream (message_start, content_block_delta, etc.).
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/models:
    get:
      summary: List available models
      description: |
        Returns the unified catalog of available models across all active upstreams,
        annotated with provider prefixes, benchmark rankings, context limits, and pricing tags.
      operationId: listModels
      tags:
        - Models
      responses:
        '200':
          description: List of available model entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/quota:
    get:
      summary: Quota status dashboard API
      description: |
        Returns the quota status across all pooled provider subscriptions so monitoring probes
        and dashboard web UIs work without modification.

        Authentication follows the rest of the surface: with client keys (or an
        admin api_key) configured the request must carry one, because the
        response names the provider lanes and their live quota state. A daemon
        started with no keys configured serves it unauthenticated, exactly like
        every other route.
      operationId: getQuotaStatus
      tags:
        - Dashboard & Quotas
      responses:
        '200':
          description: Provider quota status summary and window details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaDashboardResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/stats/summary:
    get:
      summary: System usage and accounting summary
      description: |
        Returns aggregated usage metrics read live from the daemon's SQLite telemetry store:
        request count, prompt/completion/cached token totals, the number of distinct client
        keys that produced traffic, a per-provider breakdown, and the estimated cost saved by
        pooling subscriptions.

        `estimated_cost_saved_usd` is the pay-as-you-go equivalent of the recorded traffic
        (the sum of each model's alias-priced usage cost): the pooled subscriptions are flat
        fee, so that is what the same tokens would have cost from a metered API. A daemon
        built without storage answers the full shape with zeros.
      operationId: getStatsSummary
      tags:
        - Accounting & Metrics
      responses:
        '200':
          description: Usage summary across all clients and providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsSummaryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/stats/by-client:
    get:
      summary: Per-client token accounting breakdown
      description: |
        Returns per-API-key accounting metrics, tracking which agent or developer consumed tokens,
        request counts, errors, and current rate-limiting states.

        `client_id` is the configured client key name when the stored key digest is known, and
        the digest itself otherwise (rows recorded before keys were configured, or with no key at
        all, report `anonymous`). Records are ordered by request count, descending.
      operationId: getStatsByClient
      tags:
        - Accounting & Metrics
      parameters:
        - name: window
          in: query
          required: false
          description: |
            Lookback window for the aggregation, as a Go duration or a day count
            (for example 1h, 24h, 7d, 30d). Defaults to 7d. An unparseable value
            is rejected with 400.
          schema:
            type: string
            default: 7d
            example: 7d
      responses:
        '200':
          description: Per-client usage breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientUsageRecord'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /mcp:
    get:
      summary: Streamable HTTP Model Context Protocol (MCP) endpoint
      description: |
        Streamable HTTP transport endpoint for MCP clients and AI agents.
        Exposes built-in tools for model introspection, quota queries, and dynamic model toggling.
      operationId: getMcpStream
      tags:
        - Model Context Protocol
      responses:
        '200':
          description: MCP streaming connection established
          content:
            text/event-stream:
              schema:
                type: string
            application/json:
              schema:
                type: object
    post:
      summary: Post MCP message
      description: Send JSON-RPC 2.0 messages to the embedded MCP server over HTTP transport.
      operationId: postMcpMessage
      tags:
        - Model Context Protocol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpJsonRpcRequest'
      responses:
        '200':
          description: MCP JSON-RPC response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpJsonRpcResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
      description: "Ultiproxy client key sent as Authorization: Bearer <key> (the OpenAI-compatible form). The key comes from the daemon config: the admin api_key or one of the client_keys."
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
      description: "Ultiproxy client key sent as x-api-key: <key> (the Anthropic form, so an unmodified Anthropic SDK authenticates against POST /v1/messages). The same key set as bearerAuth is accepted."
  responses:
    BadRequestError:
      description: Invalid request payload or parameter format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Missing or invalid authorization token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitError:
      description: Quota or rate limit exceeded on upstream providers
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal proxy error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGatewayError:
      description: Upstream provider failure or connection timeout
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ChatCompletionRequest:
      type: object
      required:
        - model
        - messages
      properties:
        model:
          type: string
          description: Model identifier (e.g. gpt-4o, claude-3-5-sonnet, deepseek-r1, copilot/gpt-4o)
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        stream:
          type: boolean
          default: false
        temperature:
          type: number
          minimum: 0
          maximum: 2
        top_p:
          type: number
          minimum: 0
          maximum: 1
        max_tokens:
          type: integer
        max_completion_tokens:
          type: integer
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolDefinition'
        tool_choice:
          oneOf:
            - type: string
            - type: object
        stream_options:
          type: object
          properties:
            include_usage:
              type: boolean
...
    ChatCompletionResponse:
      type: object
      required:
        - id
        - object
        - created
        - model
        - choices
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            required:
              - index
              - message
              - finish_reason
            properties:
              index:
                type: integer
              message:
                $ref: '#/components/schemas/ChatMessage'
              finish_reason:
                type: string
        usage:
          $ref: '#/components/schemas/UsageStats'
...
    AnthropicMessageRequest:
      type: object
      required:
        - model
        - messages
        - max_tokens
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            type: object
            required:
              - role
              - content
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                oneOf:
                  - type: string
                  - type: array
                    items:
                      type: object
        max_tokens:
          type: integer
        system:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
        temperature:
          type: number
        stream:
          type: boolean
        tools:
          type: array
          items:
            type: object
    AnthropicMessageResponse:
      type: object
      required:
        - id
        - type
        - role
        - model
        - content
      properties:
        id:
          type: string
        type:
          type: string
          enum: [message]
        role:
          type: string
          enum: [assistant]
        model:
          type: string
        content:
          type: array
          items:
            type: object
            required:
              - type
            properties:
              type:
                type: string
                enum: [text, tool_use]
              text:
                type: string
        stop_reason:
          type: string
        usage:
          type: object
          properties:
            input_tokens:
              type: integer
            output_tokens:
              type: integer

    # --- Models Catalog Schema ---
    ModelListResponse:
      type: object
      required:
        - object
        - data
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: '#/components/schemas/ModelRecord'
...
    QuotaDashboardResponse:
      type: object
      required:
        - providers
        - summary
      properties:
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderSummary'
        summary:
          $ref: '#/components/schemas/QuotaSummary'
...
    StatsSummaryResponse:
      type: object
      required:
        - total_requests
        - total_prompt_tokens
        - total_completion_tokens
        - active_clients
        - estimated_cost_saved_usd
        - provider_breakdown
      properties:
        total_requests:
          type: integer
          description: Terminal request rows recorded in the telemetry store.
        total_prompt_tokens:
          type: integer
        total_completion_tokens:
          type: integer
        total_cached_tokens:
          type: integer
        active_clients:
          type: integer
          description: Distinct client key hashes that produced at least one request.
        estimated_cost_saved_usd:
          type: number
          description: |
            Pay-as-you-go equivalent of the recorded traffic (the sum of the
            alias-priced usage cost); the pooled subscriptions are flat fee, so
            this is what the same tokens would have cost from a metered API.
        provider_breakdown:
          type: object
          description: Requests, tokens and terminal errors per provider lane (keyed by lane name).
          additionalProperties:
            type: object
            required:
              - requests
              - tokens
              - errors
            properties:
              requests:
                type: integer
              tokens:
                type: integer
              errors:
                type: integer
        total_tokens:
          type: integer
          deprecated: true
          description: Prompt + completion tokens; kept for the pre-contract payload.
        total_cost:
          type: number
          deprecated: true
          description: Sum of the recorded usage cost; same value as estimated_cost_saved_usd.
    ClientUsageRecord:
      type: object
      required:
        - client_id
        - request_count
        - prompt_tokens
        - completion_tokens
        - last_active
      properties:
        client_id:
          type: string
          description: Configured client key name, or the key digest when unknown, or "anonymous".
        client_name:
          type: string
          description: Present only when the digest maps to a configured client key name.
        client_key_hash:
          type: string
          description: sha256 hex digest of the client key as recorded in telemetry.
        request_count:
          type: integer
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        cached_tokens:
          type: integer
        total_tokens:
          type: integer
        cost:
          type: number
          description: Alias-priced USD cost of the window's traffic.
        error_count:
          type: integer
          description: Requests that terminated with an error class.
        rate_limited_count:
          type: integer
          description: Requests whose error class is a rate limit.
        last_active:
          type: string
          format: date-time

    # --- Health & Discovery Schemas ---
...
    McpJsonRpcRequest:
      type: object
      required:
        - jsonrpc
        - method
        - id
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        method:
          type: string
          enum:
            - tools/list
            - tools/call
            - resources/list
            - resources/read
            - prompts/list
            - initialize
        id:
          oneOf:
            - type: string
            - type: integer
        params:
          type: object
...